requirements.txt
numpy
pandas
scipy
matplotlib
seaborn
plotly
scikit-learn
joblib
catboost
statsmodels
sqlalchemy
psycopg2-binary
requests
python-dotenv
flask
fastapi
uvicorn
pydantic
streamlit
streamlit-autorefresh
jupyter
ipykernel
gpxpy
geopandas
shapely
pyproj
folium
tqdm